requirements.txt
fastapi>=0.122.0
uvicorn[standard]>=0.38.0
sqlalchemy>=2.0.44
pydantic>=2.11.7
pydantic-settings>=2.12.0
email-validator>=2.3.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0
python-multipart>=0.0.20
jinja2>=3.1.6
aiofiles>=25.1.0
requests>=2.31.0

